+++ src/fare-contracts/use-validity-line-colors.ts
@@ -1,0 +1,24 @@
+import {useFirestoreConfiguration} from '@atb/configuration';
+import {useTheme} from '@atb/theme';
+import {useTransportationColor} from '@atb/utils/use-transportation-color';
+
+export function useValidityLineColors(fareProductType?: string) {
+  const {theme} = useTheme();
+  const {fareProductTypeConfigs} = useFirestoreConfiguration();
+
+  const fareProductTypeConfig = fareProductTypeConfigs.find(
+    (c) => c.type === fareProductType,
+  );
+
+  const {mode, subMode} = fareProductTypeConfig?.transportModes?.[0] || {};
+
+  let lineColor = theme.static.background.background_accent_0.background;
+  const backgroundColor = useTransportationColor(mode, subMode);
+
+  return {
+    lineColor,
+    backgroundColor: mode
+      ? backgroundColor
+      : theme.static.status.valid.background,
+  };
+}
